requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.13.1
    # via flask-migrate
aniso8601==10.0.1
    # via flask-restful
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
bcrypt==4.1.2
    # via -r requirements.in
blinker==1.9.0
    # via flask
cachetools==5.5.2
    # via google-auth
certifi==2025.6.15
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via flask
colorama==0.4.6
    # via
    #   click
    #   pytest
deprecated==1.2.18
    # via limits
dnspython==2.7.0
    # via email-validator
email-validator==2.1.0
    # via -r requirements.in
flasgger==0.9.7.1
    # via -r requirements.in
flask==3.0.0
    # via
    #   -r requirements.in
    #   flasgger
    #   flask-cors
    #   flask-jwt-extended
    #   flask-limiter
    #   flask-migrate
    #   flask-restful
    #   flask-sqlalchemy
flask-cors==4.0.0
    # via -r requirements.in
flask-jwt-extended==4.6.0
    # via -r requirements.in
flask-limiter==3.12
    # via -r requirements.in
flask-migrate==4.0.7
    # via -r requirements.in
flask-restful==0.3.10
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via
    #   -r requirements.in
    #   flask-migrate
google-api-core==2.25.1
    # via google-api-python-client
google-api-python-client==2.174.0
    # via -r requirements.in
google-auth==2.27.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-auth-oauthlib==1.2.0
    # via -r requirements.in
googleapis-common-protos==1.70.0
    # via google-api-core
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.10
    # via
    #   email-validator
    #   requests
iniconfig==2.1.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
jsonschema==4.24.0
    # via flasgger
jsonschema-specifications==2025.4.1
    # via jsonschema
limits==5.4.0
    # via flask-limiter
mako==1.3.10
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
mistune==3.1.3
    # via flasgger
oauthlib==3.3.1
    # via requests-oauthlib
ordered-set==4.1.0
    # via flask-limiter
packaging==25.0
    # via
    #   flasgger
    #   limits
    #   pytest
pluggy==1.6.0
    # via pytest
proto-plus==1.26.1
    # via google-api-core
protobuf==6.31.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pygments==2.19.2
    # via rich
pyjwt==2.10.1
    # via flask-jwt-extended
pymysql==1.1.0
    # via -r requirements.in
pyparsing==3.2.3
    # via httplib2
pytest==8.0.0
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2025.2
    # via flask-restful
pyyaml==6.0.2
    # via flasgger
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rich==13.9.4
    # via flask-limiter
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via
    #   flasgger
    #   flask-restful
sqlalchemy==2.0.35
    # via
    #   alembic
    #   flask-sqlalchemy
typing-extensions==4.14.0
    # via
    #   alembic
    #   limits
    #   sqlalchemy
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
werkzeug==3.0.1
    # via
    #   flask
    #   flask-jwt-extended
wrapt==1.17.2
    # via deprecated
